+++ src/pages/Overview/normalizeLiveNow.js
